requirements/dev.txt
-r base.txt
django-extensions
pipdeptree
pydot
black # code formatting
django-debug-toolbar
django-cors-headers
